add pdf format

+++ twain-scanner/sources/TwainWeb.Standalone/App/Controllers/HomeController.cs
@@ -46,8 +46,8 @@ namespace TwainWeb.Standalone.App.Controllers
         }
 
         public ActionResult DownloadFile(DownloadFileParam fileParam)
-        {   
-            foreach(var fileCheck in fileParam.ListFiles)
+        {
+            foreach (var fileCheck in fileParam.ListFiles)
             {
                 if (!GlobalDictionaries.Scans.Exists(x => x == fileCheck.TempFile))
                     throw new Exception("Не найден файл для скачивания!");
@@ -57,20 +57,23 @@ namespace TwainWeb.Standalone.App.Controllers
             var fileId = fileParam.ListFiles[0].TempFile;
             var fileName = fileParam.ListFiles[0].FileName;
             string mimeType = null;
-            
-	        if (fileParam.SaveAs == (int) GlobalDictionaries.SaveAsValues.Pdf)
-	        {
-		        fileId = MakePdf(fileParam.ListFiles);
-		        mimeType = _mimeTypes["pdf"];
-				fileName = Path.ChangeExtension(fileName, ".pdf");
-	        }
-	        else
-	        {
-				var extension = Path.GetExtension(fileName);
-				if (!string.IsNullOrEmpty(extension))
-					mimeType = _mimeTypes[extension.Substring(1).ToLower()];
-	        }
+
+            if (fileParam.SaveAs == (int)GlobalDictionaries.SaveAsValues.Pdf)
+            {
+                fileId = MakePdf(fileParam.ListFiles);
+                mimeType = _mimeTypes["pdf"];
+                fileName = Path.ChangeExtension(fileName, ".pdf");
+            }
+            else
+            {
+                var extension = Path.GetExtension(fileName);
+                if (!string.IsNullOrEmpty(extension))
+                    mimeType = _mimeTypes[extension.Substring(1).ToLower()];
+            }
             var file = Path.Combine(tempDir, fileId);
+
+            string base64str = "";
+
             if (File.Exists(file))
             {
                 response = File.ReadAllBytes(file);
@@ -80,22 +83,84 @@ namespace TwainWeb.Standalone.App.Controllers
                 {
                     if (File.Exists(tempDir + img.TempFile))
                     {
-						File.Delete(tempDir + img.TempFile);
+                        File.Delete(tempDir + img.TempFile);
                         GlobalDictionaries.Scans.Remove(img.TempFile);
                     }
                 }
             }
             else
                 response = Encoding.UTF8.GetBytes("Not found");
-            return new ActionResult { Content = response, ContentType = mimeType, FileNameToDownload = fileName };
+            
+
+            //отправляем не файл а base64
+            return
+                new ActionResult { Content = response, ContentType = mimeType, FileNameToDownload = fileName };
         }
+
+        public ActionResult ConvertToBase64File(DownloadFileParam fileParam)
+        {
+            foreach (var fileCheck in fileParam.ListFiles)
+            {
+                if (!GlobalDictionaries.Scans.Exists(x => x == fileCheck.TempFile))
+                    throw new Exception("Не найден файл для скачивания!");
+            }
+            var tempDir = Path.GetTempPath();
+            byte[] response;
+            var fileId = fileParam.ListFiles[0].TempFile;
+            var fileName = fileParam.ListFiles[0].FileName;
+            string mimeType = null;
+
+            if (fileParam.SaveAs == (int)GlobalDictionaries.SaveAsValues.Pdf)
+            {
+                fileId = MakePdf(fileParam.ListFiles);
+                mimeType = _mimeTypes["pdf"];
+                fileName = Path.ChangeExtension(fileName, ".pdf");
+            }
+            else
+            {
+                var extension = Path.GetExtension(fileName);
+                if (!string.IsNullOrEmpty(extension))
+                    mimeType = _mimeTypes[extension.Substring(1).ToLower()];
+            }
+            var file = Path.Combine(tempDir, fileId);
+
+            string base64str = "";
+
+            if (File.Exists(file))
+            {
+                response = File.ReadAllBytes(file);
+                //массив байт конвертируем в строку
+                base64str = Convert.ToBase64String(response);
+                File.Delete(file);
+                GlobalDictionaries.Scans.Remove(fileId);
+                foreach (var img in fileParam.ListFiles)
+                {
+                    if (File.Exists(tempDir + img.TempFile))
+                    {
+                        File.Delete(tempDir + img.TempFile);
+                        GlobalDictionaries.Scans.Remove(img.TempFile);
+                    }
+                }
+            }
+            else
+                response = Encoding.UTF8.GetBytes("Not found");
+
+
+            var singleScanResult = new SingleScanResult();
+            singleScanResult.FillContent(new DownloadFile(fileName, fileId, base64str));
+
+            //отправляем не файл а base64
+            return                
+                new ActionResult { Content = singleScanResult.Content, ContentType = "text/json" };
+        }
+
         private string MakePdf(IEnumerable<DownloadFile> listPictures)
         {
             try
             {
                 using (var document = new PdfDocument())
                 {
-                    document.Info.Creator = "TWAIN@Web 1.3beta (http://unit6.ru/twain-web)";
+                    document.Info.Creator = "iGov";
                     var dir = Path.GetTempPath();
                     foreach (var image in listPictures)
                     {
